backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlmodel
psycopg2-binary
passlib
pyjwt
bcrypt